refactor(swingset): split test-device-mailbox.js to a separate file
no behavioral changes, although the two tests now run in parallel whereas
before they were serialized (probably unnecessarily)
